another piece of battles
Yifeng, your blitting optimizations are good enough for now (and work for me ;]). If you want to do something simple and developing, you could write sound handling - especially a function to play sounds directly from lod files (atm it requires unpacking to a file and opening it). It may require a small patch to SDL_mixer, but will significantly contribute to this project.

git-svn-id: https://vcmi.svn.sourceforge.net/svnroot/vcmi/trunk@396 4dfc6b2a-8b31-0410-aa91-d6c70e3ac834